requirements.txt
discord.py==1.5
appdirs==1.4
